@@ -32,11 +27,9 @@ authors = [
 maintainers = [
     {name = 'MDAnalysis Core Developers', email = 'mdanalysis@numfocus.org'}
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    'numpy>=1.22.3',
-    'biopython>=1.80',
-    'networkx>=2.0',
+    'numpy>=1.23.2',
     'GridDataFormats>=0.4.0',
     'mmtf-python>=1.0.0',
     'joblib>=0.12',
