@@ -987,6 +987,7 @@ class ProtoReader(IOBase, metaclass=_Readermeta):
         ))
 
     def timeseries(self, asel: Optional['AtomGroup']=None,
+                   atomgroup: Optional['Atomgroup']=None,
                    start: Optional[int]=None, stop: Optional[int]=None,
                    step: Optional[int]=None,
                    order: Optional[str]='fac') -> np.ndarray:
@@ -998,6 +999,12 @@ class ProtoReader(IOBase, metaclass=_Readermeta):
             The :class:`~MDAnalysis.core.groups.AtomGroup` to read the
             coordinates from. Defaults to ``None``, in which case the full set
             of coordinate data is returned.
+
+            .. deprecated:: 2.7.0
+                asel argument will be renamed to atomgroup in 3.0.0
+
+        atomgroup: AtomGroup (optional)
+            Same as `asel`, will replace `asel` in 3.0.0
         start :  int (optional)
             Begin reading the trajectory at frame index `start` (where 0 is the
             index of the first frame in the trajectory); the default
@@ -1023,14 +1030,22 @@ class ProtoReader(IOBase, metaclass=_Readermeta):
 
         .. versionadded:: 2.4.0
         """
+        if asel is not None:
+            warnings.warn(
+                "asel argument to timeseries will be renamed to"
+                "'atomgroup' in 3.0, see #3911",
+                category=DeprecationWarning)
+            if atomgroup:
+                raise ValueError("Cannot provide both asel and atomgroup kwargs")
+            atomgroup = asel
         start, stop, step = self.check_slice_indices(start, stop, step)
         nframes = len(range(start, stop, step))
 
-        if asel is not None:
-            if len(asel) == 0:
+        if atomgroup is not None:
+            if len(atomgroup) == 0:
                 raise ValueError(
                     "Timeseries requires at least one atom to analyze")
-            atom_numbers = asel.indices
+            atom_numbers = atomgroup.indices
             natoms = len(atom_numbers)
         else:
             natoms = self.n_atoms
