@@ -27,7 +27,7 @@ import pytest
 
 import numpy as np
 
-from numpy.testing import assert_equal, assert_almost_equal
+from numpy.testing import assert_equal, assert_allclose
 
 import MDAnalysis as mda
 from MDAnalysis.analysis import base
@@ -179,6 +179,11 @@ def u():
     return mda.Universe(PSF, DCD)
 
 
+@pytest.fixture(scope='module')
+def u_xtc():
+    return mda.Universe(TPR, XTC)  # dt = 100
+
+
 FRAMES_ERR = 'AnalysisBase.frames is incorrect'
 TIMES_ERR = 'AnalysisBase.times is incorrect'
 
@@ -194,7 +199,7 @@ def test_start_stop_step(u, run_kwargs, frames):
     assert an.n_frames == len(frames)
     assert_equal(an.found_frames, frames)
     assert_equal(an.frames, frames, err_msg=FRAMES_ERR)
-    assert_almost_equal(an.times, frames+1, decimal=4, err_msg=TIMES_ERR)
+    assert_allclose(an.times, frames+1, rtol=0, atol=1.5e-4, err_msg=TIMES_ERR)
 
 
 @pytest.mark.parametrize('run_kwargs, frames', [
@@ -205,9 +210,8 @@ def test_start_stop_step(u, run_kwargs, frames):
     ({'frames': [True, True, False, True, False, True, True, False, True,
                  False]}, (0, 1, 3, 5, 6, 8)),
 ])
-def test_frame_slice(run_kwargs, frames):
-    u = mda.Universe(TPR, XTC)  # dt = 100
-    an = FrameAnalysis(u.trajectory).run(**run_kwargs)
+def test_frame_slice(u_xtc, run_kwargs, frames):
+    an = FrameAnalysis(u_xtc.trajectory).run(**run_kwargs)
     assert an.n_frames == len(frames)
     assert_equal(an.found_frames, frames)
     assert_equal(an.frames, frames, err_msg=FRAMES_ERR)
@@ -229,29 +233,26 @@ def test_frame_fail(u, run_kwargs):
         an.run(**run_kwargs)
 
 
-def test_frame_bool_fail():
-    u = mda.Universe(TPR, XTC)  # dt = 100
-    an = FrameAnalysis(u.trajectory)
+def test_frame_bool_fail(u_xtc):
+    an = FrameAnalysis(u_xtc.trajectory)
     frames = [True, True, False]
-    msg = 'boolean index did not match indexed array along dimension 0'
+    msg = 'boolean index did not match indexed array along (axis|dimension) 0'
     with pytest.raises(IndexError, match=msg):
         an.run(frames=frames)
 
 
-def test_rewind():
-    u = mda.Universe(TPR, XTC)  # dt = 100
-    an = FrameAnalysis(u.trajectory).run(frames=[0, 2, 3, 5, 9])
-    assert_equal(u.trajectory.ts.frame, 0)
+def test_rewind(u_xtc):
+    FrameAnalysis(u_xtc.trajectory).run(frames=[0, 2, 3, 5, 9])
+    assert_equal(u_xtc.trajectory.ts.frame, 0)
 
 
-def test_frames_times():
-    u = mda.Universe(TPR, XTC)  # dt = 100
-    an = FrameAnalysis(u.trajectory).run(start=1, stop=8, step=2)
+def test_frames_times(u_xtc):
+    an = FrameAnalysis(u_xtc.trajectory).run(start=1, stop=8, step=2)
     frames = np.array([1, 3, 5, 7])
     assert an.n_frames == len(frames)
     assert_equal(an.found_frames, frames)
     assert_equal(an.frames, frames, err_msg=FRAMES_ERR)
-    assert_almost_equal(an.times, frames*100, decimal=4, err_msg=TIMES_ERR)
+    assert_allclose(an.times, frames*100, rtol=0, atol=1.5e-4, err_msg=TIMES_ERR)
 
 
 def test_verbose(u):
@@ -260,23 +261,24 @@ def test_verbose(u):
 
 
 def test_verbose_progressbar(u, capsys):
-    an = FrameAnalysis(u.trajectory).run()
-    out, err = capsys.readouterr()
+    FrameAnalysis(u.trajectory).run()
+    _, err = capsys.readouterr()
     expected = ''
     actual = err.strip().split('\r')[-1]
     assert actual == expected
 
 
 def test_verbose_progressbar_run(u, capsys):
-    an = FrameAnalysis(u.trajectory).run(verbose=True)
-    out, err = capsys.readouterr()
+    FrameAnalysis(u.trajectory).run(verbose=True)
+    _, err = capsys.readouterr()
     expected = u'100%|██████████| 98/98 [00:00<00:00, 8799.49it/s]'
     actual = err.strip().split('\r')[-1]
     assert actual[:24] == expected[:24]
 
 def test_verbose_progressbar_run_with_kwargs(u, capsys):
-    an = FrameAnalysis(u.trajectory).run(verbose=True, progressbar_kwargs={'desc':'custom'})
-    out, err = capsys.readouterr()
+    FrameAnalysis(u.trajectory).run(
+        verbose=True, progressbar_kwargs={'desc': 'custom'})
+    _, err = capsys.readouterr()
     expected = u'custom: 100%|██████████| 98/98 [00:00<00:00, 8799.49it/s]'
     actual = err.strip().split('\r')[-1]
     assert actual[:30] == expected[:30]
@@ -372,7 +374,7 @@ def test_AnalysisFromFunction_args_content(u):
     ans = base.AnalysisFromFunction(mass_xyz, protein, another, masses)
     assert len(ans.args) == 3
     result = np.sum(ans.run().results.timeseries)
-    assert_almost_equal(result, -317054.67757345125, decimal=6)
+    assert_allclose(result, -317054.67757345125, rtol=0, atol=1.5e-6)
     assert (ans.args[0] is protein) and (ans.args[1] is another)
     assert ans._trajectory is protein.universe.trajectory
 
