@@ -21,7 +21,7 @@
 # J. Comput. Chem. 32 (2011), 2319--2327, doi:10.1002/jcc.21787
 #
 import numpy as np
-from numpy.testing import assert_almost_equal, assert_equal
+from numpy.testing import assert_equal, assert_allclose
 import matplotlib
 import pytest
 
@@ -46,7 +46,7 @@ class TestDihedral(object):
         dihedral = Dihedral([atomgroup]).run()
         test_dihedral = np.load(DihedralArray)
 
-        assert_almost_equal(dihedral.results.angles, test_dihedral, 5,
+        assert_allclose(dihedral.results.angles, test_dihedral, rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -54,7 +54,7 @@ class TestDihedral(object):
         dihedral = Dihedral([atomgroup]).run(start=5, stop=6)
         test_dihedral = [np.load(DihedralArray)[5]]
 
-        assert_almost_equal(dihedral.results.angles, test_dihedral, 5,
+        assert_allclose(dihedral.results.angles, test_dihedral, rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test vales")
 
@@ -62,7 +62,7 @@ class TestDihedral(object):
         dihedral = Dihedral([atomgroup, atomgroup]).run()
         test_dihedral = np.load(DihedralsArray)
 
-        assert_almost_equal(dihedral.results.angles, test_dihedral, 5,
+        assert_allclose(dihedral.results.angles, test_dihedral, rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -91,7 +91,7 @@ class TestRamachandran(object):
     def test_ramachandran(self, universe, rama_ref_array):
         rama = Ramachandran(universe.select_atoms("protein")).run()
 
-        assert_almost_equal(rama.results.angles, rama_ref_array, 5,
+        assert_allclose(rama.results.angles, rama_ref_array, rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -99,7 +99,7 @@ class TestRamachandran(object):
         rama = Ramachandran(universe.select_atoms("protein")).run(
             start=5, stop=6)
 
-        assert_almost_equal(rama.results.angles[0], rama_ref_array[5], 5,
+        assert_allclose(rama.results.angles[0], rama_ref_array[5], rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -107,7 +107,7 @@ class TestRamachandran(object):
         rama = Ramachandran(universe.select_atoms("resname GLY")).run()
         test_rama = np.load(GLYRamaArray)
 
-        assert_almost_equal(rama.results.angles, test_rama, 5,
+        assert_allclose(rama.results.angles, test_rama, rtol=0, atol=1.5e-5,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -169,14 +169,14 @@ class TestJanin(object):
         janin = Janin(u.select_atoms("protein")).run()
 
         # Test precision lowered to account for platform differences with osx
-        assert_almost_equal(janin.results.angles, ref_array, 3,
+        assert_allclose(janin.results.angles, ref_array, rtol=0, atol=1.5e-3,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
     def test_janin_single_frame(self, universe, janin_ref_array):
         janin = Janin(universe.select_atoms("protein")).run(start=5, stop=6)
 
-        assert_almost_equal(janin.results.angles[0], janin_ref_array[5], 3,
+        assert_allclose(janin.results.angles[0], janin_ref_array[5], rtol=0, atol=1.5e-3,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
@@ -184,7 +184,7 @@ class TestJanin(object):
         janin = Janin(universe.select_atoms("resname LYS")).run()
         test_janin = np.load(LYSJaninArray)
 
-        assert_almost_equal(janin.results.angles, test_janin, 3,
+        assert_allclose(janin.results.angles, test_janin, rtol=0, atol=1.5e-3,
                             err_msg="error: dihedral angles should "
                             "match test values")
 
