@@ -25,6 +25,7 @@ import textwrap
 from io import StringIO
 import itertools
 import numpy as np
+from numpy.lib import NumpyVersion
 from numpy.testing import(
     assert_equal,
 )
@@ -539,7 +540,10 @@ class TestSelectionsTPR(object):
 
 class TestSelectionRDKit(object):
     def setup_class(self):
-        pytest.importorskip("rdkit.Chem")
+        if NumpyVersion(np.__version__) < "2.0.0":
+            pytest.importorskip("rdkit.Chem")
+        else:
+            pytest.skip("RDKit does not support NumPy 2")
 
     @pytest.fixture
     def u(self):
@@ -1413,13 +1417,19 @@ def test_negative_resid():
     ("aromaticity False", 15),
 ])
 def test_bool_sel(selstr, n_atoms):
-    pytest.importorskip("rdkit.Chem")
+    if NumpyVersion(np.__version__) >= "2.0.0":
+        pytest.skip("RDKit does not support NumPy 2")
+    else:
+        pytest.importorskip("rdkit.Chem")
     u = MDAnalysis.Universe.from_smiles("Nc1cc(C[C@H]([O-])C=O)c[nH]1")
     assert len(u.select_atoms(selstr)) == n_atoms
 
 
 def test_bool_sel_error():
-    pytest.importorskip("rdkit.Chem")
+    if NumpyVersion(np.__version__) >= "2.0.0":
+        pytest.skip("RDKit does not support NumPy 2")
+    else:
+        pytest.importorskip("rdkit.Chem")
     u = MDAnalysis.Universe.from_smiles("Nc1cc(C[C@H]([O-])C=O)c[nH]1")
     with pytest.raises(SelectionError, match="'fragrant' is an invalid value"):
         u.select_atoms("aromaticity fragrant")
