@@ -47,6 +47,7 @@ __all__ = [
     "PSF_NAMD_GBIS", "DCD_NAMD_GBIS",  # NAMD, implicit solvent, 100 steps,  #1819
     "PSF_nosegid",  # psf without a segid, Issue 121
     "PSF_cmap",  # ala3 PSF from ParmEd test files with cmap
+    "PSF_inscode",  # PSF file with insertion codes
     "PDB_small",  # PDB
     "PDB_closed",
     "PDB_multiframe",
@@ -82,14 +83,14 @@ __all__ = [
     "TPR450", "TPR451", "TPR452", "TPR453", "TPR454", "TPR455", "TPR455Double",
     "TPR460", "TPR461", "TPR502", "TPR504", "TPR505", "TPR510", "TPR2016",
     "TPR2018", "TPR2019B3", "TPR2020B2", "TPR2020", "TPR2020Double",
-    "TPR2021", "TPR2021Double", "TPR2022RC1", "TPR2023",
+    "TPR2021", "TPR2021Double", "TPR2022RC1", "TPR2023", "TPR2024",
     "TPR510_bonded", "TPR2016_bonded", "TPR2018_bonded", "TPR2019B3_bonded",
     "TPR2020B2_bonded", "TPR2020_bonded", "TPR2020_double_bonded",
     "TPR2021_bonded", "TPR2021_double_bonded", "TPR2022RC1_bonded",
-    "TPR334_bonded", "TPR2023_bonded",
+    "TPR334_bonded", "TPR2023_bonded", "TPR2024_bonded",
     "TPR_EXTRA_2021", "TPR_EXTRA_2020", "TPR_EXTRA_2018",
     "TPR_EXTRA_2016", "TPR_EXTRA_407", "TPR_EXTRA_2022RC1",
-    "TPR_EXTRA_2023",
+    "TPR_EXTRA_2023", "TPR_EXTRA_2024",
     "PDB_sub_sol", "PDB_sub_dry",  # TRRReader sub selection
     "TRR_sub_sol",
     "XTC_sub_sol",
@@ -123,6 +124,7 @@ __all__ = [
     "PQR_icodes",  # PQR v2 with icodes
     "PDBQT_input",  # PDBQT
     "PDBQT_querypdb",
+    "PDBQT_tyrosol",
     "PDB_multipole",
     "FASTA",  # sequence alignment, Issue 112 + 113
     "HELANAL_BENDING_MATRIX",  # HELANAL test (from PSF+DCD (AdK) helix 8)
@@ -155,6 +157,7 @@ __all__ = [
     "LAMMPSdata_deletedatoms",  # with deleted atoms
     "LAMMPSdata_triclinic",  # lammpsdata file to test triclinic dimension parsing, albite with most atoms deleted
     "LAMMPSdata_PairIJ",  # lammps datafile with a PairIJ Coeffs section
+    "LAMMPSdata_additional_columns",  # structure for the additional column lammpstrj
     "LAMMPSDUMP",
     "LAMMPSDUMP_long",  # lammpsdump file with a few zeros sprinkled in the first column first frame
     "LAMMPSDUMP_allcoords",  # lammpsdump file with all coordinate conventions (x,xs,xu,xsu) present, from LAMMPS rdf example
@@ -165,6 +168,7 @@ __all__ = [
     "LAMMPSDUMP_chain1", # Lammps dump file with chain reader
     "LAMMPSDUMP_chain2", # Lammps dump file with chain reader
     "LAMMPS_chain", # Lammps data file with chain reader
+    "LAMMPSDUMP_additional_columns",  # lammpsdump file with additional data (an additional charge column)
     "unordered_res",  # pdb file with resids non sequential
     "GMS_ASYMOPT",  # GAMESS C1  optimization
     "GMS_SYMOPT",   # GAMESS D4h optimization
@@ -210,6 +214,12 @@ __all__ = [
     "legacy_DCD_NAMD_coords",  # frame 0 read in for SiN_tric_namd.dcd using legacy DCD reader
     "legacy_DCD_c36_coords",  # frames 1 and 4 read in for tip125_tric_C36.dcd using legacy DCD reader
     "GSD", "GSD_bonds", "GSD_long",
+    "TRC_PDB_VAC", "TRC_TRAJ1_VAC", "TRC_TRAJ2_VAC",  # 2x 3 frames of vacuum trajectory from GROMOS11 tutorial
+    "TRC_CLUSTER_VAC",  # three frames without TIMESTEP and GENBOX block but with unsupported POSITION block
+    "TRC_TRICLINIC_SOLV", "TRC_TRUNCOCT_VAC",
+    "TRC_GENBOX_ORIGIN", "TRC_GENBOX_EULER",
+    "TRC_EMPTY",  # Empty file containing only one space
+    "TRC_PDB_SOLV", "TRC_TRAJ_SOLV",  # 2 frames of solvated trajectory from GROMOS11 tutorial
     "GRO_MEMPROT", "XTC_MEMPROT",  # YiiP transporter in POPE:POPG lipids with Na+, Cl-, Zn2+ dummy model without water
     "DihedralArray", "DihedralsArray",  # time series of single dihedral
     "RamaArray", "GLYRamaArray",  # time series of phi/psi angles
@@ -241,6 +251,7 @@ __all__ = [
     "PDB_charges",  # PDB file with formal charges
     "SURFACE_PDB",  # 111 FCC lattice topology for NSGrid bug #2345
     "SURFACE_TRR",  # full precision coordinates for NSGrid bug #2345
+    "DSSP",  # DSSP test suite
 ]
 
 from importlib import resources
@@ -304,6 +315,8 @@ DCD_NAMD_GBIS = (_data_ref / 'adk_gbis_tmd-fast1_NAMD.dcd').as_posix()
 PSF_nosegid = (_data_ref / 'nosegid.psf').as_posix()
 
 PSF_cmap = (_data_ref / 'parmed_ala3.psf').as_posix()
+
+PSF_inscode = (_data_ref / '1a2c_ins_code.psf').as_posix()
 
 PDB_varying = (_data_ref / 'varying_occ_tmp.pdb').as_posix()
 PDB_small = (_data_ref / 'adk_open.pdb').as_posix()
@@ -389,6 +402,7 @@ TPR2020 = (_data_ref / 'tprs/2lyz_gmx_2020.tpr').as_posix()
 TPR2021 = (_data_ref / 'tprs/2lyz_gmx_2021.tpr').as_posix()
 TPR2022RC1 = (_data_ref / 'tprs/2lyz_gmx_2022-rc1.tpr').as_posix()
 TPR2023 = (_data_ref / 'tprs/2lyz_gmx_2023.tpr').as_posix()
+TPR2024 = (_data_ref / 'tprs/2lyz_gmx_2024.tpr').as_posix()
 # double precision
 TPR455Double = (_data_ref / 'tprs/drew_gmx_4.5.5.double.tpr').as_posix()
 TPR460 = (_data_ref / 'tprs/ab42_gmx_4.6.tpr').as_posix()
@@ -408,7 +422,9 @@ TPR2021_bonded = (_data_ref / 'tprs/all_bonded/dummy_2021.tpr').as_posix()
 TPR2021_double_bonded = (_data_ref / 'tprs/all_bonded/dummy_2021_double.tpr').as_posix()
 TPR2022RC1_bonded = (_data_ref / 'tprs/all_bonded/dummy_2022-rc1.tpr').as_posix()
 TPR2023_bonded = (_data_ref / 'tprs/all_bonded/dummy_2023.tpr').as_posix()
+TPR2024_bonded = (_data_ref / 'tprs/all_bonded/dummy_2024.tpr').as_posix()
 # all interactions
+TPR_EXTRA_2024 = (_data_ref / 'tprs/virtual_sites/extra-interactions-2024.tpr').as_posix()
 TPR_EXTRA_2023 = (_data_ref / 'tprs/virtual_sites/extra-interactions-2023.tpr').as_posix()
 TPR_EXTRA_2022RC1 = (_data_ref / 'tprs/virtual_sites/extra-interactions-2022-rc1.tpr').as_posix()
 TPR_EXTRA_2021 = (_data_ref / 'tprs/virtual_sites/extra-interactions-2021.tpr').as_posix()
@@ -474,8 +490,9 @@ PRM19SBOPC = (_data_ref / 'Amber/ala.ff19SB.OPC.parm7.bz2').as_posix()
 PQR = (_data_ref / 'adk_open.pqr').as_posix()
 PQR_icodes = (_data_ref / '1A2C.pqr').as_posix()
 
-PDBQT_input = (_data_ref / 'pdbqt_inputpdbqt.pdbqt').as_posix()
-PDBQT_querypdb = (_data_ref / 'pdbqt_querypdb.pdb').as_posix()
+PDBQT_input = (_data_ref / "pdbqt_inputpdbqt.pdbqt").as_posix()
+PDBQT_querypdb = (_data_ref / "pdbqt_querypdb.pdb").as_posix()
+PDBQT_tyrosol = (_data_ref / "tyrosol.pdbqt.bz2").as_posix()
 
 FASTA = (_data_ref / 'test.fasta').as_posix()
 HELANAL_BENDING_MATRIX = (_data_ref / 'helanal_bending_matrix_AdK_DIMS_H8.dat').as_posix()
@@ -544,6 +561,8 @@ LAMMPSDUMP_chain1 = (_data_ref / "lammps/chain_dump_1.lammpstrj").as_posix()
 LAMMPSDUMP_chain2 = (_data_ref / "lammps/chain_dump_2.lammpstrj").as_posix()
 LAMMPS_chain = (_data_ref / "lammps/chain_initial.data").as_posix()
 LAMMPSdata_many_bonds = (_data_ref / "lammps/a_lot_of_bond_types.data").as_posix()
+LAMMPSdata_additional_columns = (_data_ref / "lammps/additional_columns.data").as_posix()
+LAMMPSDUMP_additional_columns = (_data_ref / "lammps/additional_columns.lammpstrj").as_posix()
 
 unordered_res = (_data_ref / "unordered_res.pdb").as_posix()
 
@@ -597,6 +616,18 @@ GSD = (_data_ref / 'example.gsd').as_posix()
 GSD_bonds = (_data_ref / 'example_bonds.gsd').as_posix()
 GSD_long = (_data_ref / 'example_longer.gsd').as_posix()
 
+TRC_PDB_VAC = (_data_ref / 'gromos11/gromos11_traj_vac.pdb.gz').as_posix()
+TRC_TRAJ1_VAC = (_data_ref / 'gromos11/gromos11_traj_vac_1.trc.gz').as_posix()
+TRC_TRAJ2_VAC = (_data_ref / 'gromos11/gromos11_traj_vac_2.trc.gz').as_posix()
+TRC_PDB_SOLV = (_data_ref / 'gromos11/gromos11_traj_solv.pdb.gz').as_posix()
+TRC_TRAJ_SOLV = (_data_ref / 'gromos11/gromos11_traj_solv.trc.gz').as_posix()
+TRC_CLUSTER_VAC = (_data_ref / 'gromos11/gromos11_cluster_vac.trj.gz').as_posix()
+TRC_TRICLINIC_SOLV = (_data_ref / 'gromos11/gromos11_triclinic_solv.trc.gz').as_posix()
+TRC_TRUNCOCT_VAC = (_data_ref / 'gromos11/gromos11_truncOcta_vac.trc.gz').as_posix()
+TRC_GENBOX_ORIGIN = (_data_ref / 'gromos11/gromos11_genbox_origin.trc.gz').as_posix()
+TRC_GENBOX_EULER = (_data_ref / 'gromos11/gromos11_genbox_euler.trc.gz').as_posix()
+TRC_EMPTY = (_data_ref / 'gromos11/gromos11_empty.trc').as_posix()
+
 DihedralArray = (_data_ref / 'adk_oplsaa_dihedral.npy').as_posix()
 DihedralsArray = (_data_ref / 'adk_oplsaa_dihedral_list.npy').as_posix()
 RamaArray = (_data_ref / 'adk_oplsaa_rama.npy').as_posix()
@@ -632,6 +663,9 @@ PDBX = (_data_ref / "4x8u.pdbx").as_posix()
 
 SURFACE_PDB = (_data_ref / 'surface.pdb.bz2').as_posix()
 SURFACE_TRR = (_data_ref / 'surface.trr').as_posix()
+
+# DSSP testing: from https://github.com/ShintaroMinami/PyDSSP
+DSSP = (_data_ref / 'dssp').as_posix()
 
 # This should be the last line: clean up namespace
 del resources
